Refactored IO handlers
